Scratch code for related files
After a bit of experimenting, it seems as if transforms will cause a
lot of complexity, due to say, when visiting foo_test.cc, needing to
strip off _test from the filename, and having this be something the
user specifies in the transforms is not simple in the general
case (i.e. for controller, model, view, etc).  It seems simpler to
have a list of suffixes that are generally present in files that are
related to each other and strip that suffix from the currently-visited
file, if present, before generating a list of filenames that are
related.